pyproject.toml
[project]
name = "rss-summary"
version = "0.1.0"
description = "Summarize RSS feeds using Claude AI"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "anthropic",
    "feedparser",
    "python-dateutil",
]
